requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.2.0=pypi_0
appnope=0.1.2=py310hecd8cb5_1001
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=openblas
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.4.26=hecd8cb5_0
cachetools=5.2.0=pypi_0
cairo=1.16.0=h691a603_2
certifi=2022.6.15=py310hecd8cb5_0
charset-normalizer=2.1.1=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py310he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.3.7=haf03e11_1
entrypoints=0.4=py310hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
flatbuffers=2.0.7=pypi_0
fontconfig=2.13.1=ha9ee91d_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
gast=0.4.0=pypi_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=h8346a28_1
google-auth=2.11.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=he9d5cce_1
grpcio=1.48.1=pypi_0
gst-plugins-base=1.14.0=h4180768_2
gstreamer=1.14.0=h0fc69c2_2
h5py=3.7.0=pypi_0
harfbuzz=4.3.0=h3f129a2_0
hdf5=1.10.6=hdbbcd12_0
icu=58.2=h0a44026_3
idna=3.3=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.1=py310hecd8cb5_0
ipython=8.4.0=py310hecd8cb5_0
jedi=0.18.1=py310hecd8cb5_1
jpeg=9e=hca72f7f_0
jupyter_client=7.2.2=py310hecd8cb5_0
jupyter_core=4.10.0=py310hecd8cb5_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=py310he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=14.0.6=pypi_0
libcxx=12.0.0=h2f01273_0
libedit=3.1.20210910=hca72f7f_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=hca72f7f_2
libllvm12=12.0.0=h9b2ccf5_3
libopenblas=0.3.20=h9a5756b_0
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_3
libsodium=1.0.18=h1de35cc_0
libtiff=4.2.0=hdb42f99_1
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
llvm-openmp=12.0.0=h0dcd299_1
lxml=4.9.1=py310h65b224f_0
lz4-c=1.9.3=h23ab428_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=py310hecd8cb5_1
matplotlib-base=3.5.1=py310hfb0c5b7_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.5=py310hecd8cb5_0
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numexpr=2.8.3=py310h6c44a92_0
numpy=1.21.5=py310h6c44a92_3
numpy-base=1.21.5=py310h8a30af7_3
oauthlib=3.2.0=pypi_0
opencv=4.5.5=py310h86893f8_4
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1o=hca72f7f_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310he9d5cce_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h23ab428_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py310hde71d04_0
pip=21.2.4=py310hecd8cb5_0
pixman=0.40.0=h9ed2024_1
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
python=3.10.4=hdfd78df_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.1=py310hecd8cb5_0
pyzmq=23.2.0=py310he9d5cce_0
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_4
qtwebkit=5.212=h24dc246_4
readline=8.1.2=hca72f7f_1
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.7.3=py310hbd46d07_0
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=61.2.0=py310hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h707629a_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
tornado=6.1=py310hca72f7f_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=pypi_0
tzdata=2022a=hda174b7_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xz=5.2.5=hca72f7f_1
zeromq=4.3.4=h23ab428_0
zlib=1.2.12=h4dc903c_2
zstd=1.5.2=hcb37349_0
